third_party/WebKit/LayoutTests/fast/html/imports/import-cycle-grandparent.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../http/tests/htmlimports/resources/normalize-test-helper.js"></script>
<script>
var t = async_test('Imports with cycle complete loading in a defined order.');
window.addEventListener('load', function() {
    t.step(function() {
        assert_array_equals(['cycle-grandparent-c.html', 'cycle-grandparent-b.html', 'cycle-grandparent-a.html'], importedDocumentList);
        t.done();
    });
});
</script>
</head>
<body>
<link rel="import" href="resources/cycle-grandparent-a.html">
</body>
</html>
